response to minecraft link/unlink commands

+++ cogs/minecraft_link.py
@@ -1,5 +1,6 @@
 from discord import *
 from data_class import MinecraftMemberData
+from utils.bot_embeds import SucceedEmbed, DangerEmbed, InformativeEmbed
 
 class MinecraftLinkCog(Cog):
     def __init__(self, bot):
@@ -13,10 +14,17 @@ class MinecraftLinkCog(Cog):
         minecraft_member_data = MinecraftMemberData(ctx.author, ctx.guild)
         await minecraft_member_data.set_uuid_from_name(username)
 
+        await ctx.respond(embed=SucceedEmbed(title="Compté lié", description=f"Ton compte discord a été lié au compte minecraft `{username}`, ton pseudo sera mis-à-jour dans quelques minutes"))
+
     @minecraft.command(name="unlink")
     async def minecraft_unlink(self, ctx):
         minecraft_member_data = MinecraftMemberData(ctx.author, ctx.guild)
-        minecraft_member_data.delete()
+        if minecraft_member_data.file_exist:
+            minecraft_member_data.delete()
+            await ctx.respond(embed=DangerEmbed(title="Compte délié", description="Ton compte discord n'est désormais plus lié"))
+        else:
+            await ctx.respond(embed=InformativeEmbed(title="Aucun lien existant", description="Ton compte discord n'est pas lié"))
+
 
 
 def setup(bot):
